REFACTOR: mock 데이터 분리
